requirements.txt
opencv-python~=4.10.0.84
numpy~=1.26.0
torch~=2.5.1
segmentation-models-pytorch~=0.5.0
tqdm~=4.67.1
pyyaml~=6.0.2
pillow~=11.1.0
matplotlib~=3.10.0
scikit-learn~=1.7.1
pyzed~=4.2
ultralytics~=8.0.196